+++ src/main/java/edu/hbuas/controller/FixedController.java
@@ -4,6 +4,8 @@ import edu.hbuas.common.ResponseJson;
 import edu.hbuas.pojo.Fixed;
 import edu.hbuas.service.FixedService;
 import edu.hbuas.vo.FindView;
+import org.apache.shiro.SecurityUtils;
+import org.apache.shiro.subject.Subject;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -24,18 +26,30 @@ public class FixedController {
     @RequestMapping(value = "entry.do", method = RequestMethod.POST)
     @ResponseBody
     public ResponseJson<String> entry(String cardId) {
+        Subject subject = SecurityUtils.getSubject();
+        if(subject.hasRole("user")&&!subject.isPermitted("fixed:entry")) {
+            return ResponseJson.createByErrorNoPer();
+        }
         return fixedService.insertFixed(cardId);
     }
 
     @RequestMapping(value = "out.do", method = RequestMethod.POST)
     @ResponseBody
     public ResponseJson<String> out(String fixedId) {
+        Subject subject = SecurityUtils.getSubject();
+        if(subject.hasRole("user")&&!subject.isPermitted("fixed:out")) {
+            return ResponseJson.createByErrorNoPer();
+        }
         return fixedService.insertFixedOut(fixedId);
     }
 
     @RequestMapping(value = "delete.do", method = RequestMethod.POST)
     @ResponseBody
     public ResponseJson<String> delete(String fixedId) {
+        Subject subject = SecurityUtils.getSubject();
+        if(subject.hasRole("user")&&!subject.isPermitted("fixed:delete")) {
+            return ResponseJson.createByErrorNoPer();
+        }
         return fixedService.deleteFixed(fixedId);
     }
 
@@ -43,6 +57,10 @@ public class FixedController {
     @RequestMapping(value = "find.do", method = RequestMethod.POST)
     @ResponseBody
     public ResponseJson find(@RequestBody FindView findFixedView){
+        Subject subject = SecurityUtils.getSubject();
+        if(subject.hasRole("user")&&!subject.isPermitted("fixed:find")) {
+            return ResponseJson.createByErrorNoPer();
+        }
         System.out.println(findFixedView);
         return fixedService.selectAll(findFixedView);
     }
@@ -50,6 +68,10 @@ public class FixedController {
     @RequestMapping(value = "findEntry.do", method = RequestMethod.POST)
     @ResponseBody
     public ResponseJson findEntry(@RequestBody FindView findFixedView){
+        Subject subject = SecurityUtils.getSubject();
+        if(subject.hasRole("user")&&!subject.isPermitted("fixed:findEntry")) {
+            return ResponseJson.createByErrorNoPer();
+        }
         System.out.println(findFixedView);
         return fixedService.selectAllEntry(findFixedView);
     }
